+++ .storybook/main.js
@@ -6,6 +6,7 @@ module.exports = {
     '../packages/styled-components/components/**/*.stories.(js|mdx)',
   ],
   addons: [
+    '@storybook/addon-a11y',
     '@storybook/addon-actions',
     '@storybook/addon-docs',
     '@storybook/addon-links',
@@ -16,6 +17,11 @@ module.exports = {
         WEBPACK_BUILD: true,
       })
     );
+    config.module.rules[0].exclude.push([
+      /\bcore-js\b/,
+      /\bwebpack\/buildin\b/
+    ]);
+
     return config;
   },
 };