+++ packages/core/config/aliases.js
@@ -15,6 +15,7 @@ module.exports = {
   server: './server',
   services: './services',
   utils: './utils',
+  config: './config',
   // Aliases for irving config files. Use core defaults if they don't exist.
   '@irvingjs/irving.config': maybeResolveBuildModule('irving.config.js'),
   '@irvingjs/irving.config.server': maybeResolveBuildModule(