+++ packages/core/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.3](https://github.com/alleyinteractive/irving/packages/core/compare/v2.21.2...v2.21.3) (2020-07-07)
+
+
+### Bug Fixes
+
+* **storybook:** add componentMap to storybook to prevent errors ([9da288f](https://github.com/alleyinteractive/irving/packages/core/commit/9da288f0d5316a68b82964e30a8c981762c32f32))
+
+
+
+
+
+# [2.20.0](https://github.com/alleyinteractive/irving/packages/core/compare/v2.19.3-beta.0...v2.20.0) (2020-07-01)
+
+**Note:** Version bump only for package @irvingjs/core
+
+
+
+
+
+## [2.19.3-beta.0](https://github.com/alleyinteractive/irving/packages/core/compare/v2.19.2...v2.19.3-beta.0) (2020-07-01)
+
+**Note:** Version bump only for package @irvingjs/core
+
+
+
+
+
 ## [2.19.2](https://github.com/alleyinteractive/irving/packages/core/compare/v2.19.1...v2.19.2) (2020-06-30)
 
 **Note:** Version bump only for package @irvingjs/core